re-enable selective's default consumer and add a test for it
We're also doing something with the monitors now: selective keeps
track of which consumers are still alive; when it receives a message,
it either routes it to the appropriate consumer if it was registered
and is still alive, or to the default consumer if it exists, or drops
it otherwise.  Monitors are used to remove dead consumers.

There's a race between the consumer dying and messages for it: it's
possible for the consumer to die, and a few messages to get lost
before selective receives the DOWN and starts re-routing messages to
the default consumer.