rename the server_initiated_close in channel to something more appropriate
Turns out server_initiated_channel had two distinct meetings.  A
channel closing to this reason is one that received a frame that
couldn't be processed by rabbit_command_assembler.  A connection
closing to this reason is one that received a connection.close.  So,
rename the one in channel and replace the one in connection with the
actual #'connection.close'{}.